make comment deletion more RESTful

+++ blueprints.py
@@ -165,21 +165,24 @@ def projects():
     return render_template("projects.html", header = title, title = title, user = current_user, comments = comments, form = form)
 
 # route to handle comment deletion, comment id is passed in from deleteComment.js script
-@blueprints.route("/delete_comment/<int:comment_id>", methods=["DELETE"])
+@blueprints.route("/comments/<int:comment_id>", methods=["DELETE"])
 @login_required
 def delete_comment(comment_id):
     try:
         comment = Comments.query.get(comment_id)
 
+        if not comment:
+            return "Comment not found", 404
+
         # prevents deletion of other user comments
         if current_user.id != comment.user_id:
             flash("You do not have permission to delete this comment.", category = "form_error")
-            return redirect(url_for(request.endpoint))
+            return "Permission denied", 403
 
         db.session.delete(comment)
         db.session.commit()
         flash("Comment deleted successfully.", category = "form_success")
-        return "Comment deleted successfully", 200
+        return "Comment deleted successfully", 204
     except Exception as e:
         print(f"An error occurred: {e}")
         flash("An error occurred while deleting the comment.", category = "form_error")
